Move EndMessagePersonality to own file
